+++ src/main/java/com/project/adaptationflow/entity/links/NotificationUserLinkId.java
@@ -1,4 +1,4 @@
-package com.project.adaptationflow.entity;
+package com.project.adaptationflow.entity.links;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -14,28 +14,28 @@ import java.util.UUID;
 @Getter
 @Setter
 @Embeddable
-public class UserAchievementLinkId implements Serializable {
-    private static final long serialVersionUID = 1092624886342258426L;
+public class NotificationUserLinkId implements Serializable {
+    private static final long serialVersionUID = 6895816219259468112L;
+    @NotNull
+    @Column(name = "notification_id", nullable = false)
+    private UUID notificationId;
+
     @NotNull
     @Column(name = "user_id", nullable = false)
     private UUID userId;
-
-    @NotNull
-    @Column(name = "achievement_id", nullable = false)
-    private UUID achievementId;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        UserAchievementLinkId entity = (UserAchievementLinkId) o;
-        return Objects.equals(this.achievementId, entity.achievementId) &&
+        NotificationUserLinkId entity = (NotificationUserLinkId) o;
+        return Objects.equals(this.notificationId, entity.notificationId) &&
                 Objects.equals(this.userId, entity.userId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(achievementId, userId);
+        return Objects.hash(notificationId, userId);
     }
 
 }
