+++ src/main/java/com/project/adaptationflow/entity/links/UserOnboardingTrackLink.java
@@ -1,5 +1,6 @@
-package com.project.adaptationflow.entity;
+package com.project.adaptationflow.entity.links;
 
+import com.project.adaptationflow.entity.tasks.OnboardingTrack;
 import com.project.adaptationflow.entity.user.SysUser;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -10,13 +11,10 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "user_achievement_link", indexes = {
-        @Index(name = "idx_user_achievement_link_user_id", columnList = "user_id"),
-        @Index(name = "idx_user_achievement_link_achievement_id", columnList = "achievement_id")
-})
-public class UserAchievementLink {
+@Table(name = "user_onboarding_track_link")
+public class UserOnboardingTrackLink {
     @EmbeddedId
-    private UserAchievementLinkId id;
+    private UserOnboardingTrackLinkId id;
 
     @MapsId("userId")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
@@ -24,10 +22,10 @@ public class UserAchievementLink {
     @JoinColumn(name = "user_id", nullable = false)
     private SysUser user;
 
-    @MapsId("achievementId")
+    @MapsId("trackId")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @OnDelete(action = OnDeleteAction.CASCADE)
-    @JoinColumn(name = "achievement_id", nullable = false)
-    private Achievement achievement;
+    @JoinColumn(name = "track_id", nullable = false)
+    private OnboardingTrack track;
 
 }
