+++ src/main/java/com/project/adaptationflow/entity/links/UserOnboardingTrackLinkId.java
@@ -1,4 +1,4 @@
-package com.project.adaptationflow.entity;
+package com.project.adaptationflow.entity.links;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -14,28 +14,28 @@ import java.util.UUID;
 @Getter
 @Setter
 @Embeddable
-public class NotificationUserLinkId implements Serializable {
-    private static final long serialVersionUID = 6895816219259468112L;
-    @NotNull
-    @Column(name = "notification_id", nullable = false)
-    private UUID notificationId;
-
+public class UserOnboardingTrackLinkId implements Serializable {
+    private static final long serialVersionUID = 5467964054475266876L;
     @NotNull
     @Column(name = "user_id", nullable = false)
     private UUID userId;
+
+    @NotNull
+    @Column(name = "track_id", nullable = false)
+    private UUID trackId;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        NotificationUserLinkId entity = (NotificationUserLinkId) o;
-        return Objects.equals(this.notificationId, entity.notificationId) &&
+        UserOnboardingTrackLinkId entity = (UserOnboardingTrackLinkId) o;
+        return Objects.equals(this.trackId, entity.trackId) &&
                 Objects.equals(this.userId, entity.userId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(notificationId, userId);
+        return Objects.hash(trackId, userId);
     }
 
 }
