+++ src/main/java/com/project/adaptationflow/entity/links/UsersTasksLink.java
@@ -1,5 +1,6 @@
-package com.project.adaptationflow.entity;
+package com.project.adaptationflow.entity.links;
 
+import com.project.adaptationflow.entity.tasks.Task;
 import com.project.adaptationflow.entity.user.SysUser;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -10,10 +11,13 @@ import org.hibernate.annotations.OnDeleteAction;
 @Getter
 @Setter
 @Entity
-@Table(name = "user_onboarding_track_link")
-public class UserOnboardingTrackLink {
+@Table(name = "users_tasks_link", indexes = {
+        @Index(name = "idx_users_tasks_link_user_id", columnList = "user_id"),
+        @Index(name = "idx_users_tasks_link_task_id", columnList = "task_id")
+})
+public class UsersTasksLink {
     @EmbeddedId
-    private UserOnboardingTrackLinkId id;
+    private UsersTasksLinkId id;
 
     @MapsId("userId")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
@@ -21,10 +25,10 @@ public class UserOnboardingTrackLink {
     @JoinColumn(name = "user_id", nullable = false)
     private SysUser user;
 
-    @MapsId("trackId")
+    @MapsId("taskId")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @OnDelete(action = OnDeleteAction.CASCADE)
-    @JoinColumn(name = "track_id", nullable = false)
-    private OnboardingTrack track;
+    @JoinColumn(name = "task_id", nullable = false)
+    private Task task;
 
 }
