+++ src/main/java/com/project/adaptationflow/entity/links/UsersTasksLinkId.java
@@ -1,4 +1,4 @@
-package com.project.adaptationflow.entity;
+package com.project.adaptationflow.entity.links;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -14,28 +14,28 @@ import java.util.UUID;
 @Getter
 @Setter
 @Embeddable
-public class UserOnboardingTrackLinkId implements Serializable {
-    private static final long serialVersionUID = 5467964054475266876L;
+public class UsersTasksLinkId implements Serializable {
+    private static final long serialVersionUID = -8733541137490251069L;
     @NotNull
     @Column(name = "user_id", nullable = false)
     private UUID userId;
 
     @NotNull
-    @Column(name = "track_id", nullable = false)
-    private UUID trackId;
+    @Column(name = "task_id", nullable = false)
+    private UUID taskId;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        UserOnboardingTrackLinkId entity = (UserOnboardingTrackLinkId) o;
-        return Objects.equals(this.trackId, entity.trackId) &&
-                Objects.equals(this.userId, entity.userId);
+        UsersTasksLinkId entity = (UsersTasksLinkId) o;
+        return Objects.equals(this.userId, entity.userId) &&
+                Objects.equals(this.taskId, entity.taskId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(trackId, userId);
+        return Objects.hash(userId, taskId);
     }
 
 }
